+++ utils/callDeviceGroupApi.tsx
@@ -1,3 +1,6 @@
+import { DeviceGroup } from "@/components/types";
+import { HTTPStatus } from "@/utils/utils";
+
 function isValidDeviceGroupName(name: string): boolean {
   return /^[a-zA-Z0-9-_]+$/.test(name);
 }
@@ -17,7 +20,7 @@ export const apiGetAllDeviceGroups = async (token: string) => {
   }
 };
 
-export const apiGetDeviceGroup = async (name: string, token: string) => {
+export async function apiGetDeviceGroup(name: string, token: string): Promise<Response> {
   if (!isValidDeviceGroupName(name)) {
     throw new Error(`Error getting device group: Invalid name provided ${name}.`);
   }
@@ -30,8 +33,8 @@ export const apiGetDeviceGroup = async (name: string, token: string) => {
       },
     });
   } catch (error) {
-    console.error(error);
-    throw error;
+      console.error("Unexpected error:", error);
+      throw new Error(`Unexpected error occurred: ${error}`);
   }
 };
