+++ utils/utils.ts
@@ -37,3 +37,4 @@ export const HTTPStatus = (code: number): string => {
     }
     return map[code]
 }
+
